refactor: events serializer functions moved to core
